feat(internationzalization): add multilingual web support with Thymeleaf and locale resolution
Ever since my days as a student, I've loved automatic internationzalization.
Godot and Java do it and it scratches something in my brain, so I always go
out of my way to get them into as many projects as I can.

I also preconfigured some things that I'll need in the future, like the template,
and the language changer button. I also found out what a favicon.ico is. All
thanks to the security logger, notifying me. Had to add the UTF-8 encoding property
to the properties to avoid some issues. Might change the exclamation signs to unicode
just to be safe.

Technically, this feature is done, but I want to test it out a tad more, so for now it stays
open and it's branch as well.

As per usual, here's the summary:

- Introduce `LocaleConfig` with `SessionLocaleResolver` and `LocaleChangeInterceptor`
- Add Thymeleaf templates: `index.html`, `template.html`, `langbutton.html` for UI with language switcher
- Register "/" view controller in `WebMvcConfig` to serve home page
- Add image assets for language flags and favicon (EN, ES, CA)
- Serve static resources like `/images/*` and `/favicon.ico` publicly
- Update `application.properties` to set UTF-8 encoding for messages
- Prepare `TodoRepository` for pagination (TODO comment)