refactor(compose): Refactor images for Redis and Postgres
In the same line as the previous commit/branch, a refactor of the images
inside the compose were made. While more testing will be done, from
preliminary testing, no changes are made to our current workflow since
we work with the uses that the alpine version of both Redis and Postgres
offer.

This is preliminary work, it might change in the near future.