refactor(index): Create base controller for index page
Delete WebMvcConfig, as there's no longer a need for it. All endpoints
are manually handled. Create an empty TodoWebController, as it'll be
filled promptly. Create BaseWebController, for handling the "/" and
"/index" endpoints, as some logic needs to be handled in it.

Update index.html for some thymeleaf testing.

Leaving it here for today, I've got to sleep.